These test will now pass with subpixel fix for scrollX, scrollY
Re-gen'd SDK with new Window.idl (double scrollX, scrollY, pageXOffset and pageYOffset).

Reviewers=vsm@google.com,asiva@google.com

R=asiva@google.com

Review URL: https://codereview.chromium.org//864663003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@43480 260f80e4-7a28-3924-810f-c04153c831b5

+++ sdk/lib/html/dart2js/html_dart2js.dart
@@ -30618,7 +30618,7 @@ class Window extends EventTarget implements WindowEventHandlers, WindowBase, Glo
    */
   @DomName('Window.pageXOffset')
   @DocsEditable()
-  final int _pageXOffset;
+  final double _pageXOffset;
 
   @JSName('pageYOffset')
   /**
@@ -30635,7 +30635,7 @@ class Window extends EventTarget implements WindowEventHandlers, WindowBase, Glo
    */
   @DomName('Window.pageYOffset')
   @DocsEditable()
-  final int _pageYOffset;
+  final double _pageYOffset;
 
   @DomName('Window.parent')
   @DocsEditable()
